mapped home index to site root.
