+++ src/NavigationRoutes/NavigationRoutes/NavigationRouteConfigurationExtensions.cs
@@ -1,0 +1,93 @@
+﻿using Microsoft.Web.Mvc;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Web;
+using System.Web.Mvc;
+using System.Web.Mvc.Html;
+using System.Web.Routing;
+
+namespace NavigationRoutes
+{
+    public static class NavigationRouteConfigurationExtensions
+    {
+
+        public static void MapNavigationRoute(this RouteCollection routes, string name, string url, object defaults,
+                                              object constraints = null)
+        {
+            var newRoute = new NamedRoute(name, url, new MvcRouteHandler()) { Defaults = new RouteValueDictionary(defaults), Constraints = new RouteValueDictionary(constraints) };
+            routes.Add(name, newRoute);
+        }
+
+        public static void MapNavigationRoute(this RouteCollection routes, string name, string displayName, string url, object defaults,
+                                              object constraints = null)
+        {
+            var newRoute = new NamedRoute(name, displayName, url, new MvcRouteHandler()) { Defaults = new RouteValueDictionary(defaults), Constraints = new RouteValueDictionary(constraints) };
+            routes.Add(name, newRoute);
+        }
+
+
+        public static void MapNavigationRoute<T>(this RouteCollection routes, string displayName, Expression<Func<T, ActionResult>> action) where T : IController
+        {
+            var newRoute = new NamedRoute("", "", new MvcRouteHandler());
+            newRoute.ToDefaultAction(action);
+            //newRoute.Constraints = new RouteValueDictionary(new { @namespace=typeof(T).Namespace});
+            newRoute.DisplayName = displayName;
+            routes.Add(newRoute.Name, newRoute);
+        }
+
+        public static NamedRoute ToDefaultAction<T>(this NamedRoute route, Expression<Func<T, ActionResult>> action) where T : IController
+        {
+            var body = action.Body as MethodCallExpression;
+
+            if (body == null)
+            {
+                throw new ArgumentException("Expression must be a method call");
+            }
+
+            if (body.Object != action.Parameters[0])
+            {
+                throw new ArgumentException("Method call must target lambda argument");
+            }
+
+            string actionName = body.Method.Name;
+
+            // check for ActionName attribute
+            var attributes = body.Method.GetCustomAttributes(typeof(ActionNameAttribute), false);
+            if (attributes.Length > 0)
+            {
+                var actionNameAttr = (ActionNameAttribute)attributes[0];
+                actionName = actionNameAttr.Name;
+            }
+
+            string controllerName = typeof(T).Name;
+
+            if (controllerName.EndsWith("Controller", StringComparison.OrdinalIgnoreCase))
+            {
+                controllerName = controllerName.Remove(controllerName.Length - 10, 10);
+            }
+
+            ;
+            route.Defaults = LinkBuilder.BuildParameterValuesFromExpression(body) ?? new RouteValueDictionary();
+            foreach (var pair in route.Defaults.Where(x => x.Value == null).ToList())
+                route.Defaults.Remove(pair.Key);
+
+            route.Defaults.Add("controller", controllerName);
+            route.Defaults.Add("action", actionName);
+
+            route.Url= CreateUrl(actionName,controllerName);
+            route.Name = "Navigation-" + controllerName + "-" + actionName;
+            return route;
+        }
+
+        public static string CreateUrl(string actionName, string controllerName)
+        {
+            if (controllerName.Equals("home", StringComparison.CurrentCultureIgnoreCase) && actionName.Equals("index",StringComparison.CurrentCultureIgnoreCase))
+            {
+                return "";
+            }
+            return actionName.ToLower();
+        }
+    }
+}
